refactor: API 모듈 분리 (클라이언트/서버)
서버리스 환경 안전성과 번들 최적화를 위해 api.ts를 분리:
- api/types.ts: 공통 타입 정의
- api/client.ts: axios 인스턴스 + 토큰 리프레시 (클라이언트 전용)
- api/server.ts: fetch 기반 ISR 함수들 (서버 전용)
